refactor(venues): move seat-adjust toggle into the toolbar, inspector to a row
Moves the "Adjust seats" mode toggle out of the side-column card and into
the seat editor's action bar (now persistent, before Toggle Accessible),
with the "Add seat" sub-toggle appearing alongside it once adjust mode is
on. Selection-dependent bulk actions hide while adjust mode is active.

The single-seat inspector (SeatAdjustPanel, renamed SeatAdjustInspector)
becomes a full-width row directly below the bar, shown only when a seat
is selected in adjust mode — the right column now holds just the row
geometry panel. Pure relayout: no changes to mode gating, selection,
drag, keyboard nudging, undo/redo, or save semantics.
